+++ claude_install.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-curl -sSL https://raw.githubusercontent.com/carmandale/agent-os/main/setup.sh | bash